+++ crates/rust-client/src/store/web_store/account/utils.rs
@@ -90,8 +90,8 @@ pub async fn insert_account_record(
         vault_root,
         nonce,
         committed,
-        account_seed,
         commitment,
+        account_seed,
     );
     JsFuture::from(promise).await?;